feat: add WalletRepository.java

+++ src/main/java/com/playtomic/tests/wallet/api/repository/WalletRepository.java
@@ -3,6 +3,6 @@ package com.playtomic.tests.wallet.api.repository;
 import com.playtomic.tests.wallet.api.model.Wallet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WalletRepository extends JpaRepository<Wallet, Long> {
+public interface WalletRepository extends JpaRepository<Wallet, Integer> {
 
 }
